[GH-1] Implement basic context menu handling.
This is a very basic code that must soon be refactored. ContextMenu should be a part of a bigger struct: RectangularWindow, which is to be implemented soon. Having ContextMenu as part of the RectangularWindow should solve the problem of keeping references to ContextMenu in window memory (although it's curious mechanism).